Ported Imgui tools from EOS Branch and adapted it, fixed an issue with packets being casted to a byte when they are an int, Added preprocessor constants for WINDOWS, MAC, LINUX and FreeBSD

+++ ClassLibrary1/Networking/Packets/Architecture/PacketSender.cs
@@ -18,7 +18,7 @@ namespace ONI_MP.Networking
 			using (var ms = new System.IO.MemoryStream())
 			using (var writer = new System.IO.BinaryWriter(ms))
 			{
-				writer.Write((byte)packet.Type);
+				writer.Write((int)packet.Type);
 				packet.Serialize(writer);
 				return ms.ToArray();
 			}
